requirements.txt
# Azure KQL Explorer Web Application

flask>=2.3.0
python-dotenv>=1.0.0
azure-identity>=1.15.0
azure-monitor-query>=1.2.0
openai>=1.0.0
